requirements.txt
openai
yfinance
pandas
datetime
altair
matplotlib
seaborn
streamlit-chat
statsmodels
prophet
pandas_datareader
newspaper3k
konlpy
getmac
msoffcrypto-tool
openpyxl
pygwalker
langchain
langchain_community
tiktoken
chromadb
pykrx
finance-datareader
pinecone-client
faiss-cpu
pyrfc
pyodbc
plotly
PublicDataReader
deepl
langchain_community
google-auth-oauthlib